NSWI177_linux/22/requirements.txt
Jinja2==3.1.2
MarkupSafe==2.1.1
PyYAML==6.0
roman==3.3
